graph/iterator: consolidate node iterators
